backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase
openai==1.3.7
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
